feat: Add basic authentication for login endpoint
This commit adds basic authentication for the login endpoint in the `index.ts` file. It uses the `hono/basic-auth` module to authenticate the user with a username and password. If the authentication is successful, it returns a response with the message "You are authorized". This feature enhances the security of the application.